backend/requirements.txt

flask==3.0.0
flask-cors==4.0.0
pandas==2.1.0
numpy==1.26.0
requests==2.31.0
python-dateutil==2.8.2
vectorbt==0.26.1
yfinance==0.2.36
optuna==3.5.0
ta>=0.11.0
pyarrow>=14.0.0
dhanhq>=2.0.1
python-dotenv>=1.0.0
filelock>=3.13.0
